+++ index.js
@@ -1,9 +1,4 @@
-const {inspect} = require("util");
-
+const WebServer = require("./webapp/WebServer");
 const Config = require("./config.json");
-const MastodonAPI = require("./fediverse/MastodonAPI");
 
-const api = new MastodonAPI(Config.token, {instance: "https://masto.quad.moe"});
-
-api.fetchOwner().then((x) =>console.debug(inspect(x, {colors: true}))).catch((e) => console.log(e));
-api.fetchUser(6421).then((x) =>console.debug(inspect(x, {colors: true}))).catch((e) => console.log(e));
+new WebServer(Config.webserver).start();