+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All changes to this project wll be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2018-02-01 12:14
+### Added:
+- WebServer with basic route and error handling.
+- Logging utility with support for colored console output.
+### Updated:
+- Required dependencies for logging utility.
+
 ## [0.1.0] - 2018-02-01 01:43
 ### Added:
 - Beginning of Mastodon API.